Moved application configuration
