+++ app/console/kernel.go
@@ -2,7 +2,7 @@ package console
 
 import (
 	"github.com/RobyFerro/go-web-framework/register"
-	"github.com/RobyFerro/go-web/app"
+	"github.com/RobyFerro/go-web/config"
 	"github.com/RobyFerro/go-web/service"
 )
 
@@ -17,7 +17,7 @@ var (
 	// This container is created to allow dependency injection through console commands.
 	Services = register.ServiceRegister{
 		List: []interface{}{
-			app.Configuration,
+			config.GetConfiguration,
 			service.ConnectDB,
 			service.ConnectElastic,
 			service.ConnectMongo,
