+++ register/register.go
@@ -2,10 +2,10 @@ package register
 
 import (
 	"github.com/RobyFerro/go-web-framework/register"
-	"github.com/RobyFerro/go-web/app"
 	"github.com/RobyFerro/go-web/app/console"
 	"github.com/RobyFerro/go-web/app/http/controller"
 	"github.com/RobyFerro/go-web/app/http/middleware"
+	"github.com/RobyFerro/go-web/config"
 	"github.com/RobyFerro/go-web/database/model"
 	"github.com/RobyFerro/go-web/service"
 )
@@ -33,7 +33,7 @@ var (
 	// Every service needs to be registered in the following list
 	Services = register.ServiceRegister{
 		List: []interface{}{
-			app.Configuration,
+			config.GetConfiguration,
 			service.ConnectDB,
 			service.ConnectElastic,
 			service.ConnectMongo,
