+++ service/db_connection.go
@@ -2,14 +2,14 @@ package service
 
 import (
 	"fmt"
-	"github.com/RobyFerro/go-web/app"
+	"github.com/RobyFerro/go-web/config"
 	"github.com/jinzhu/gorm"
 	_ "github.com/jinzhu/gorm/dialects/mysql"
 	"github.com/labstack/gommon/log"
 )
 
 // ConnectDB to sql database
-func ConnectDB(conf *app.Conf) *gorm.DB {
+func ConnectDB(conf *config.Conf) *gorm.DB {
 	connectionString, driver := createConnectionString(conf)
 	db, err := gorm.Open(driver, connectionString)
 
@@ -21,7 +21,7 @@ func ConnectDB(conf *app.Conf) *gorm.DB {
 }
 
 // Create string for SQL connection
-func createConnectionString(conf *app.Conf) (string, string) {
+func createConnectionString(conf *config.Conf) (string, string) {
 	return fmt.Sprintf(
 		"%s:%s@(%s:%d)/%s?charset=utf8&parseTime=True&loc=Local",
 		conf.Database.User,
