+++ service/elasticsearch.go
@@ -1,13 +1,13 @@
 package service
 
 import (
-	"github.com/RobyFerro/go-web/app"
+	"github.com/RobyFerro/go-web/config"
 	"github.com/elastic/go-elasticsearch/v8"
 	"github.com/labstack/gommon/log"
 )
 
 // ConnectElastic returns an ElasticSearch client
-func ConnectElastic(conf *app.Conf) *elasticsearch.Client {
+func ConnectElastic(conf *config.Conf) *elasticsearch.Client {
 	cfg := elasticsearch.Config{
 		Addresses: conf.Elastic.Hosts,
 	}
