+++ service/mongo.go
@@ -3,7 +3,7 @@ package service
 import (
 	"context"
 	"fmt"
-	"github.com/RobyFerro/go-web/app"
+	"github.com/RobyFerro/go-web/config"
 	"github.com/labstack/gommon/log"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
@@ -11,7 +11,7 @@ import (
 )
 
 // ConnectMongo returns a connection to MongoDB
-func ConnectMongo(conf *app.Conf) *mongo.Database {
+func ConnectMongo(conf *config.Conf) *mongo.Database {
 	client, err := mongo.NewClient(options.Client().ApplyURI(fmt.Sprintf("mongodb://%s:%d", conf.Mongo.Host, conf.Mongo.Port)))
 
 	if err != nil {
