+++ service/redis.go
@@ -2,13 +2,13 @@ package service
 
 import (
 	"fmt"
-	"github.com/RobyFerro/go-web/app"
+	"github.com/RobyFerro/go-web/config"
 	"github.com/go-redis/redis/v7"
 	"github.com/labstack/gommon/log"
 )
 
 // ConnectRedis connect to Redis
-func ConnectRedis(conf *app.Conf) *redis.Client {
+func ConnectRedis(conf *config.Conf) *redis.Client {
 	client := redis.NewClient(&redis.Options{
 		Addr:     fmt.Sprintf("%s:%d", conf.Redis.Host, conf.Redis.Port),
 		Password: conf.Redis.Password,
